RUM-12060: Move EvictingQueue.kt back to the internal module
