+++ dd-sdk-android-core/api/dd-sdk-android-core.api
@@ -623,30 +623,6 @@ public final class com/datadog/android/core/UploadWorker : androidx/work/Worker 
 public final class com/datadog/android/core/UploadWorker$Companion {
 }
 
-public final class com/datadog/android/core/collections/EvictingQueue : java/util/Queue {
-	public fun <init> (I)V
-	public synthetic fun <init> (IILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun add (Ljava/lang/Object;)Z
-	public fun addAll (Ljava/util/Collection;)Z
-	public fun clear ()V
-	public fun contains (Ljava/lang/Object;)Z
-	public fun containsAll (Ljava/util/Collection;)Z
-	public fun element ()Ljava/lang/Object;
-	public fun getSize ()I
-	public fun isEmpty ()Z
-	public fun iterator ()Ljava/util/Iterator;
-	public fun offer (Ljava/lang/Object;)Z
-	public fun peek ()Ljava/lang/Object;
-	public fun poll ()Ljava/lang/Object;
-	public fun remove ()Ljava/lang/Object;
-	public fun remove (Ljava/lang/Object;)Z
-	public fun removeAll (Ljava/util/Collection;)Z
-	public fun retainAll (Ljava/util/Collection;)Z
-	public final fun size ()I
-	public fun toArray ()[Ljava/lang/Object;
-	public fun toArray ([Ljava/lang/Object;)[Ljava/lang/Object;
-}
-
 public final class com/datadog/android/core/configuration/BackPressureMitigation : java/lang/Enum {
 	public static final field DROP_OLDEST Lcom/datadog/android/core/configuration/BackPressureMitigation;
 	public static final field IGNORE_NEWEST Lcom/datadog/android/core/configuration/BackPressureMitigation;
