+++ dd-sdk-android-internal/api/dd-sdk-android-internal.api
@@ -30,6 +30,30 @@ public final class com/datadog/android/internal/attributes/ViewScopeInstrumentat
 	public static fun values ()[Lcom/datadog/android/internal/attributes/ViewScopeInstrumentationType;
 }
 
+public final class com/datadog/android/internal/collections/EvictingQueue : java/util/Queue {
+	public fun <init> (I)V
+	public synthetic fun <init> (IILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public fun add (Ljava/lang/Object;)Z
+	public fun addAll (Ljava/util/Collection;)Z
+	public fun clear ()V
+	public fun contains (Ljava/lang/Object;)Z
+	public fun containsAll (Ljava/util/Collection;)Z
+	public fun element ()Ljava/lang/Object;
+	public fun getSize ()I
+	public fun isEmpty ()Z
+	public fun iterator ()Ljava/util/Iterator;
+	public fun offer (Ljava/lang/Object;)Z
+	public fun peek ()Ljava/lang/Object;
+	public fun poll ()Ljava/lang/Object;
+	public fun remove ()Ljava/lang/Object;
+	public fun remove (Ljava/lang/Object;)Z
+	public fun removeAll (Ljava/util/Collection;)Z
+	public fun retainAll (Ljava/util/Collection;)Z
+	public final fun size ()I
+	public fun toArray ()[Ljava/lang/Object;
+	public fun toArray ([Ljava/lang/Object;)[Ljava/lang/Object;
+}
+
 public final class com/datadog/android/internal/network/GraphQLHeaders : java/lang/Enum {
 	public static final field DD_GRAPHQL_NAME_HEADER Lcom/datadog/android/internal/network/GraphQLHeaders;
 	public static final field DD_GRAPHQL_PAYLOAD_HEADER Lcom/datadog/android/internal/network/GraphQLHeaders;
