+++ dd-sdk-android-internal/src/main/java/com/datadog/android/internal/collections/EvictingQueue.kt
@@ -3,7 +3,7 @@
  * This product includes software developed at Datadog (https://www.datadoghq.com/).
  * Copyright 2016-Present Datadog, Inc.
  */
-package com.datadog.android.core.collections
+package com.datadog.android.internal.collections
 
 import java.util.LinkedList
 import java.util.Queue
