+++ features/dd-sdk-android-rum-debug-widget/src/main/kotlin/com/datadog/android/insights/internal/widgets/ChartView.kt
@@ -11,10 +11,10 @@ import android.graphics.Paint
 import android.graphics.Typeface
 import android.util.AttributeSet
 import android.view.View
-import com.datadog.android.core.collections.EvictingQueue
 import com.datadog.android.insights.internal.extensions.color
 import com.datadog.android.insights.internal.extensions.px
 import com.datadog.android.insights.internal.extensions.round
+import com.datadog.android.internal.collections.EvictingQueue
 import com.datadog.android.rumdebugwidget.R
 import java.util.Collections.max
 import java.util.Collections.min
